포인트 오류 수정

+++ src/main/java/com/phl/cocolo/dto/MemberUpdateDTO.java
@@ -16,7 +16,7 @@ public class MemberUpdateDTO {
     private String memberPhone;
     private MultipartFile memberProfile;
     private String memberProfileName;
-    private long memberPoint;
+    private int memberPoint;
     private String memberInteresting;
     private String memberNickName;
     private String memberLevel;
